refactor(omp): split _rpc_bridge.py before SLOC gate breach
Extract callback translation and steer subscription into dedicated
modules while keeping RpcBridge as the public facade. All three files
stay under the 300 SLOC cap; existing omp adapter tests unchanged.

Closes #1961
